Codes/requirements.txt
Flask
Flask-Cors
pyttsx3
moviepy
whisper
opencv-python
tqdm
Pillow
numpy